requirements.txt
# Data manipulation and analysis
pandas>=1.5.0
numpy>=1.23.0

# Visualization
matplotlib>=3.6.0
seaborn>=0.12.0

# Image processing
Pillow>=9.0.0

# Machine learning
scikit-learn>=1.2.0

# Deep learning
torch>=2.0.0
torchvision>=0.15.0

# Computer vision models
timm>=0.9.0

# Advanced augmentations
albumentations>=1.3.0

# Experiment tracking
mlflow>=2.8.0

# Progress bars
tqdm>=4.65.0

# Jupyter (for EDA)
jupyter>=1.0.0
ipykernel>=6.0.0

# Optional: Model inspection
torchinfo>=1.8.0

# API Framework
fastapi>=0.115.0
uvicorn[standard]>=0.32.0
python-multipart>=0.0.17
pydantic-settings>=2.0.0

